python/streamlit-multipage-sharepoint-graph/requirements.txt
--extra-index-url https://test.pypi.org/simple/
pyjwt
graphlit-client
streamlit
streamlit_extras